requirements.txt
numpy
gym==0.21.0
torch
torchvision
matplotlib
opencv-python
ipykernel
mujoco-py==2.0.2.8
cython<3
dill
python-dotenv
gymnasium[other]
miniworld
pyglet==1.5.27
stable_baselines3
wandb
